fix: AwsSchedulerService#registeredScheduleView에서 templateId가 Long으로 캐스팅 되어 있지 않은 문제 해결

+++ domain/crm/src/main/kotlin/com/few/crm/support/schedule/aws/AwsSchedulerService.kt
@@ -95,7 +95,7 @@ class AwsSchedulerService(
                     taskName = payload["eventId"] as String,
                     values =
                         mapOf(
-                            "templateId" to (payload["templateId"] as Int),
+                            "templateId" to (payload["templateId"] as Int).toLong(),
                             "userIds" to (payload["userIds"] as List<Int>).map { it.toLong() },
                             "eventId" to payload["eventId"] as String,
                             "eventType" to payload["eventType"] as String,
